+++ src/main/java/insane96mcp/survivalreimagined/module/items/solarium/item/SolariumPickaxeItem.java
@@ -1,13 +1,13 @@
-package insane96mcp.survivalreimagined.module.mining.solarium.item;
+package insane96mcp.survivalreimagined.module.items.solarium.item;
 
-import insane96mcp.survivalreimagined.module.mining.solarium.Solarium;
+import insane96mcp.survivalreimagined.module.items.solarium.Solarium;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.ShovelItem;
+import net.minecraft.world.item.PickaxeItem;
 import net.minecraft.world.level.Level;
 
-public class SolariumShovelItem extends ShovelItem {
-	public SolariumShovelItem(float pAttackDamageModifier, float pAttackSpeedModifier, Properties pProperties) {
+public class SolariumPickaxeItem extends PickaxeItem {
+	public SolariumPickaxeItem(int pAttackDamageModifier, float pAttackSpeedModifier, Properties pProperties) {
 		super(Solarium.ITEM_TIER, pAttackDamageModifier, pAttackSpeedModifier, pProperties);
 	}