+++ src/main/java/insane96mcp/survivalreimagined/module/items/solarium/item/SolariumShield.java
@@ -1,16 +1,18 @@
-package insane96mcp.survivalreimagined.module.mining.solarium.item;
+package insane96mcp.survivalreimagined.module.items.solarium.item;
 
 import insane96mcp.shieldsplus.setup.SPItems;
 import insane96mcp.shieldsplus.world.item.SPShieldItem;
 import insane96mcp.shieldsplus.world.item.SPShieldMaterial;
-import insane96mcp.survivalreimagined.module.mining.solarium.Solarium;
+import insane96mcp.survivalreimagined.module.items.solarium.Solarium;
 import insane96mcp.survivalreimagined.setup.SRRegistries;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Rarity;
 import net.minecraft.world.level.Level;
 import net.minecraftforge.registries.RegistryObject;
+import org.jetbrains.annotations.Nullable;
 
 public class SolariumShield extends SPShieldItem {
 	public static final SPShieldMaterial SHIELD_MATERIAL = new SPShieldMaterial("solarium", 4.5d, 401, Solarium.SOLARIUM_BALL, 8, Rarity.COMMON);
@@ -26,6 +28,14 @@ public class SolariumShield extends SPShieldItem {
 
 	public boolean shouldCauseReequipAnimation(ItemStack oldStack, ItemStack newStack, boolean slotChanged) {
 		return !ItemStack.isSameItem(oldStack, newStack);
+	}
+
+	@Override
+	public double getBlockedDamage(ItemStack stack, @Nullable LivingEntity entity, Level level) {
+		if (entity == null)
+			return super.getBlockedDamage(stack, null, level);
+		float calculatedLight = Solarium.getCalculatedSkyLightRatio(entity);
+		return super.getBlockedDamage(stack, entity, level) + (2f * calculatedLight);
 	}
 
 	public static RegistryObject<SPShieldItem> registerShield(String id) {